Re-enabling SwiftShader/Reactor warnings
I've found no occurrence of warning 4530 being triggered and
only 3 conversion warning 4244, so I fixed them and re-enabled
them.

+++ src/Reactor/Routine.cpp
@@ -74,7 +74,7 @@ namespace sw
 
 	int Routine::getCodeSize()
 	{
-		return functionSize - ((uintptr_t)entry - (uintptr_t)buffer);
+		return functionSize - static_cast<int>((uintptr_t)entry - (uintptr_t)buffer);
 	}
 
 	bool Routine::isDynamic()
